+++ s3-backed-core/src/main/java/com/bakdata/kafka/AbstractS3BackedConfig.java
@@ -107,14 +107,14 @@ public class AbstractS3BackedConfig extends AbstractConfig {
                 .define(ID_GENERATOR_CONFIG, Type.CLASS, ID_GENERATOR_DEFAULT, Importance.HIGH, ID_GENERATOR_DOC);
     }
 
-    S3RetrievingClient getS3Retriever() {
+    S3BackedRetrievingClient getS3Retriever() {
         final AmazonS3 s3 = this.createS3Client();
-        return new S3RetrievingClient(s3);
+        return new S3BackedRetrievingClient(s3);
     }
 
-    S3StoringClient getS3Storer() {
+    S3BackedStoringClient getS3Storer() {
         final AmazonS3 s3 = this.createS3Client();
-        return S3StoringClient.builder()
+        return S3BackedStoringClient.builder()
                 .s3(s3)
                 .basePath(this.getBasePath())
                 .maxSize(this.getMaxSize())